Add ability to add context to a Target.
By adding context a multi-stage attack can be carried out.
The context for the `Target` is passed through to `Result` which
allows you to execute another attack based on the additional context
inside the `Result`.

For example, if you want to first create a session in the first
HTTP request and then use that session ID in the second request:

```go
for res := range loginAttacker.Attack(targeter, rate, duration) {
    sessionIdChan <- res.Context["sessionId"].(string)
}
```

Then the other `Targeter` can read the `sessionId` for the next request
in the stage:

```go
return func(tgt *vegeta.Target) error {
    sessionId := <- sessionIdChan
}
```